requirements.txt
# Requirements for CMUQ Arena - Arcade Fighting Game
# Just pygame - nothing else needed
pygame-ce